fix: 将marker方法提至util，及调整其相关依赖
