+++ src/components/Overlay/Symbol.js
@@ -1,10 +1,10 @@
-import { createSymbol, getSize } from '../_base/util';
+import BaseOverlay from './BaseOverlay';
+import { createSymbol, getSize, createMarker } from '../_base/util';
 import ReactComponent from '../ReactComponent';
-import Marker from './Marker';
 
 @ReactComponent
-class Symbol extends Marker {
-  constructor(props) {
+class Symbol extends BaseOverlay {
+  init() {
     const {
       path,
       anchor = {
@@ -19,7 +19,7 @@ class Symbol extends Marker {
       strokeOpacity,
       strokeWeight,
       ...markerProps
-    } = props;
+    } = this.props;
 
     markerProps.icon = createSymbol({
       path,
@@ -35,7 +35,12 @@ class Symbol extends Marker {
       },
     });
 
-    super(markerProps);
+    this.instance = createMarker(markerProps);
+    this.map.addOverlay(this.instance);
+
+    if (markerProps.animation) {
+      this.instance.setAnimation(global[markerProps.animation]);
+    }
   }
 }
